requirements.txt
# Core dependencies
matplotlib==3.6.2
opencv-python==4.6.0.66
openslide-python==1.2.0
pandas==1.5.2
PyQt5==5.15.10
qt-material==2.14
qdarkstyle>=3.1
scipy==1.13.0
tqdm==4.66.2
multiprocess==0.70.16
orjson==3.10.0
Flask==3.0.3
scikit-image>=0.23.0
numpy>=1.26.0,<2.0.0
pillow>=9.0.0
h5py>=3.0.0
PyYAML>=6.0

# TensorFlow framework (default)
tensorflow>=2.15.0
keras>=2.15.0
segmentation-models==1.0.1

# PyTorch framework (alternative)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0

# Framework-specific dependencies
efficientnet-pytorch==0.7.1
pretrainedmodels==0.7.4
timm==0.9.2

# Testing dependencies (optional)
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
